requirements.txt
Django
djangorestframework
markdown
django-filter
django-hosts
django-rest-auth
django-allauth